Remove truncation from the semantics
In R, if a vector needs to be shorter, it gets truncated, but a warning
is issued.

Here, we decided to make such a scenario an error. Vectors can be
modified via recycling to become longer, but the final length must be a
multiple of the initial length.

Note that if the initial length is longer than the final length, it is
not a multiple of the final length.